setup_google/requirements.txt
gspread>=5.12.0
oauth2client>=4.1.3
google-api-python-client>=2.108.0
python-dotenv>=1.0.0
pytz>=2023.3 